backend-django/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
dynamic-rest==2.3.0
filelock==3.14.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
hashids==1.3.1
idna==3.7
inflection==0.5.1
PySocks==1.7.1
requests==2.31.0
requests-file==2.0.0
six==1.16.0
sqlparse==0.5.0
tldextract==5.1.2
typing_extensions==4.11.0
urllib3==2.2.1
zope.event==5.0
zope.interface==6.4
